Rust rewrite
To compile make you have rust installed, if you don't go to: https://www.rust-lang.org/tools/install/
and install it from there. To compile cd into the nvs directory and run:
cargo build --release
If you don't include --release then the program will be slower (It doesn't optimize it so it compiles faster for when your testing so when you want to bring this to production you use --release if you're just testing you don't need it) so be sure to include --release! After building go to the target directory and then from there go into the release directory and the binary named "nvs" is the binary that you're looking for.
To build and run just run:
cargo run --release